Add admin payments components & styles
Introduce a modular payments UI: add multiple React components under src/components/admin/payments (PaymentCreateModal, PaymentMarkPaidModal, PaymentStatusBadge, PaymentStatusModal, PaymentViewModal, PaymentsStats, PaymentsSummaryBar, PaymentsTable, PaymentsTabs, PaymentsToolbar) and paymentsHelpers.js (constants and formatters). Add AdminPayments.css and update src/pages/admin/Payments.jsx to import the new components and stylesheet, replacing the previous inline styles and in-file subcomponents. New modals integrate with API endpoints for creating, marking paid, and changing status; table, toolbar and tabs provide filtering, pagination and action hooks.
